Store 'testdata.json' in a directory prefixed with analyzer_assigned_uuid
Previously the prefix was a timestamp which is unreliable. This also makes
the directory prefix (the last directory name actually) a very convenient
"Result id" that is printed after each test. It somewhat streamlines
the experience between the dashboard and abrek where the UUID can
be used to lookup test runs.