clear out test_results array for each time parser is run
The test_results array is a static member of the class. This
means if you are to run the same parser twice that the results
will continue to be added to the last run's results. This is
important because an upcoming patch will allow you to run a
command like "parse" on multiple test id's and combine the results
into one bundle