glmark2: Update regexp pattern.
This commit changes the regexp pattern in two ways. First, it includes the
scene name (the part in square brackets) in the test_case_id. Second,
it removes the EOL ('$') match to make the regexp more tolerant of changes
in the output format, in this case the addition of "FrameTime" after "FPS".