Re-factor blackbox implementation
This patch changes how blackbox is integrated into LAVA. To keep this short (it
is described in great detail in the source code) due to limiations of l-a-t
run() and parse() APIs that test difinitions can provide _and_ the non-goal of
refactor the whole framework I had to monkey patch how the top-level run
command (implementation of lava-android-test run) works. This makes l-a-t work
properly with blackbox test which is my current primary goal.