cts.py: fix problem of collecting duplicate result
when adb connection becomes unavailable during the cts test,
addition to the cts output, the content of host.log file will
be output too for debug. But since the similar output is in
the host.log, the output of host.log content caused the same
test to be counted again, and as a result, we can see the cts
number doubled like here:
https://validation.linaro.org/dashboard/streams/private/team/wg/wg-private/bundles/e8d382dfe698ff41b5e9258ecc1e2fd7348da662/fc7b313d-9cac-45c1-bb2e-46f23ad3aa42/?length=25&search=android.acceleration.cts.HardwareAccelerationTesttestIsHardwareAccelerated#table

you can search android.acceleration.cts.HardwareAccelerationTesttestIsHardwareAccelerated
there, and you will see 2 entries.

And this change is to fix this problem by restricting the pattern
to only check lines that started with cts-tf prompt.

Change-Id: I75fdceb819b4a2c4f119009c16803bc26c180d49
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Reviewed-on: https://review.linaro.org/2482
Reviewed-by: Dave Pigott <dave.pigott@linaro.org>